refactor: Transaction 객체 제거
