+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -4,9 +4,7 @@ import com.techcourse.dao.UserDao;
 import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
 import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
 import org.springframework.dao.DataAccessException;
-import org.springframework.transaction.Transaction;
 import org.springframework.transaction.TransactionManager;
 
 public class UserService {
@@ -26,32 +24,29 @@ public class UserService {
     }
 
     public void insert(User user) {
-        Transaction transaction = transactionManager.getTransaction();
-        transaction.start();
+        transactionManager.start();
 
         try {
-            userDao.insert(transaction.getConnection(), user);
-            transaction.commit();
+            userDao.insert(user);
+            transactionManager.commit();
         } catch (DataAccessException e) {
-            transaction.rollback();
+            transactionManager.rollback();
             throw e;
         }
     }
 
     public void changePassword(long id, String newPassword, String createBy) {
-        Transaction transaction = transactionManager.getTransaction();
-        transaction.start();
+        transactionManager.start();
         try {
-            Connection connection = transaction.getConnection();
             User user = findById(id);
 
             user.changePassword(newPassword);
-            userDao.update(connection, user);
-            userHistoryDao.log(connection, new UserHistory(user, createBy));
+            userDao.update(user);
+            userHistoryDao.log(new UserHistory(user, createBy));
 
-            transaction.commit();
+            transactionManager.commit();
         } catch (DataAccessException e) {
-            transaction.rollback();
+            transactionManager.rollback();
             throw e;
         }
     }
