+++ jdbc/src/main/java/org/springframework/transaction/TransactionManager.java
@@ -1,23 +1,38 @@
 package org.springframework.transaction;
 
-import java.sql.SQLException;
+import java.sql.Connection;
 import javax.sql.DataSource;
-import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.datasource.DataSourceUtils;
 
 public class TransactionManager {
 
+    private final TransactionTemplate transactionTemplate;
     private final DataSource dataSource;
 
     public TransactionManager(DataSource dataSource) {
+        this.transactionTemplate = new TransactionTemplate(dataSource);
         this.dataSource = dataSource;
     }
 
-    public Transaction getTransaction() {
-        try {
-            return new Transaction(dataSource.getConnection());
-        } catch (SQLException e) {
-            throw new DataAccessException(e);
-        }
+    public void start() {
+        transactionTemplate.execute(connection -> connection.setAutoCommit(false));
+    }
+
+    public void commit() {
+        transactionTemplate.execute(Connection::commit);
+        close();
+    }
+
+    private void close() {
+        transactionTemplate.execute(connection -> {
+            connection.close();
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        });
+    }
+
+    public void rollback() {
+        transactionTemplate.execute(Connection::rollback);
+        close();
     }
 
 }
