+++ jdbc/src/main/java/org/springframework/transaction/TransactionTemplate.java
@@ -2,26 +2,26 @@ package org.springframework.transaction;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import javax.sql.DataSource;
 import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.datasource.DataSourceUtils;
 
 public class TransactionTemplate {
 
-    private final Connection connection;
+    private final DataSource dataSource;
 
-    public TransactionTemplate(Connection connection) {
-        this.connection = connection;
+    public TransactionTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
     }
 
     public void execute(TransactionExecutor executor) {
         try {
+            Connection connection = DataSourceUtils.getConnection(dataSource);
+
             executor.execute(connection);
         } catch (SQLException e) {
             throw new DataAccessException(e);
         }
-    }
-
-    public Connection getConnection() {
-        return connection;
     }
 
 }
